+++ src/Classes/Util.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * Geodata for Contao Open Source CMS
- * Copyright (c) 2023-2023 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-geodata
@@ -18,13 +18,24 @@ use Contao\Database;
 use Contao\Input;
 use Contao\StringUtil;
 use Contao\System;
+use WEM\GeoDataBundle\Model\Category;
 use WEM\GeoDataBundle\Model\MapItem;
+use WEM\GeoDataBundle\Model\MapItemCategory;
 
 /**
  * Provide utilities function to Locations Extension.
  */
 class Util
 {
+    /**
+     * Format string value for use in filters (for better readability in URL)
+     * @param  string $value The raw value
+     * @return string the formatted value
+     */
+    public static function formatStringValueForFilters(string $value): string
+    {
+        return str_replace([' ', '.'], '_', mb_strtolower($value, 'UTF-8'));
+    }
     /**
      * Calculates the great-circle distance between two points, with
      * the Vincenty formula.
@@ -399,5 +410,47 @@ class Util
         ];
 
         return $COUNTRY_CONTINENTS[strtoupper($strCountry)];
+    }
+
+    /**
+     * Delete MapItemCategory rows for a Category
+     * @param  Category $objItem The Category
+     */
+    public static function deleteMapItemCategoryForCategory(Category $objItem): void
+    {
+        // remove links item <-> category
+        $mapItemCategories = MapItemCategory::findItems(['category'=>$objItem->id]);
+        if($mapItemCategories){
+            while($mapItemCategories->next()){
+                $mapItemCategories->current()->delete();                
+            }
+        }
+    }
+    
+    /**
+     * Refreshes "categories" field for a MapItem
+     * @param  MapItem $objItem The MapItem
+     * @param  null|array $arrCategoriesIdsToExclude Ids of Category to avoid
+     * @return MapItem          The updated MapItem
+     */
+    public static function refreshMapItemCategoriesField(MapItem $objItem, ?array $arrCategoriesIdsToExclude): MapItem
+    {
+        $params = ['pid'=>$objItem->id];
+
+        if(is_array($arrCategoriesIdsToExclude)){
+            $params['where'][] = sprintf('%s.category NOT IN (%s)', MapItemCategory::getTable(), implode(',',$arrCategoriesIdsToExclude));
+        }
+
+        $mapItemCategories = MapItemCategory::findItems($params);
+        $arrCategoriesIds = [];
+        if($mapItemCategories){
+            while($mapItemCategories->next()){
+                $arrCategoriesIds[] = $mapItemCategories->category;
+            }
+        }
+        $objItem->categories = serialize($arrCategoriesIds);
+        $objItem->save();
+
+        return $objItem;
     }
 }