+++ src/DataContainer/Map.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * Geodata for Contao Open Source CMS
- * Copyright (c) 2023-2023 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-geodata
@@ -14,18 +14,48 @@ declare(strict_types=1);
 
 namespace WEM\GeoDataBundle\DataContainer;
 
-use Contao\Backend;
+use Contao\DataContainer;
+use Contao\Message;
+use WEM\GeoDataBundle\Model\Category;
 use WEM\GeoDataBundle\Model\Map as ModelMap;
 
-class Map extends Backend
+class Map extends CoreContainer
 {
-    /**
-     * Import the back end user object.
-     */
-    public function __construct()
+    public function onloadCallback(DataContainer $dc): void
     {
-        parent::__construct();
-        $this->import('BackendUser', 'User');
+        if (!$dc->id) {
+            return;
+        }
+        // check if another category is the default one for the map
+        // if not, show an error
+        $defaultCategory = Category::findItems(['pid' => $dc->id, 'is_default' => '1'], 1);
+        if (!$defaultCategory) {
+            Message::addError('No default category on this map. Add one !');
+        }
+    }
+
+    public function onsubmitCallback(DataContainer $dc): void
+    {
+        if (!$dc->id) {
+            return;
+        }
+
+        // check if another category is the default one for the map
+        // if not, make this one the default's one, sorry not sorry
+        $defaultCategory = Category::findItems(['pid' => $dc->id, 'is_default' => '1'], 1);
+        if (!$defaultCategory) {
+            $newDefaultCategory = Category::findItems(['pid' => $dc->id], 1);
+            if (!$newDefaultCategory) {
+                $newDefaultCategory = new Category();
+                $newDefaultCategory->createdAt = time();
+                $newDefaultCategory->tstamp = time();
+                $newDefaultCategory->title = 'Default';
+                $newDefaultCategory->markerConfig = serialize([]);
+                $newDefaultCategory->pid = $dc->id;
+            }
+            $newDefaultCategory->is_default = 1;
+            $newDefaultCategory->save();
+        }
     }
 
     /**