+++ src/DataContainer/MapItem.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * Geodata for Contao Open Source CMS
- * Copyright (c) 2023-2023 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-geodata
@@ -14,7 +14,6 @@ declare(strict_types=1);
 
 namespace WEM\GeoDataBundle\DataContainer;
 
-use Contao\Backend;
 use Contao\DataContainer;
 use Contao\Image;
 use Contao\Input;
@@ -22,9 +21,11 @@ use Contao\PageModel;
 use Contao\StringUtil;
 use Contao\System;
 use Contao\Versions;
+use WEM\GeoDataBundle\Model\Category;
 use WEM\GeoDataBundle\Model\Map;
+use WEM\GeoDataBundle\Model\MapItem as MapItemModel;
 
-class MapItem extends Backend
+class MapItem extends CoreContainer
 {
     /**
      * Import the back end user object.
@@ -116,6 +117,18 @@ class MapItem extends Backend
         if ('' === $objMap->geocodingProvider) {
             unset($GLOBALS['TL_DCA']['tl_wem_map_item']['list']['global_operations']['geocodeAll'], $GLOBALS['TL_DCA']['tl_wem_map_item']['list']['operations']['geocode']);
         }
+    }
+
+    public function assignDefaultCategoryIfNew($value, DataContainer $dc)
+    {
+        if (!$dc->id || !$dc->activeRecord->categories) {
+            $objDefaultCategory = Category::findItems(['pid' => $dc->activeRecord->pid, 'is_default' => '1']);
+            if ($objDefaultCategory) {
+                return serialize([$objDefaultCategory->id]);
+            }
+        }
+
+        return $value;
     }
 
     /**
@@ -285,5 +298,12 @@ class MapItem extends Backend
         $url = str_replace('&amp;id='.$objMap->id, '&amp;id='.$data['id'], $url);
 
         return sprintf('<a href="%s" title="%s" %s>%s</a> ', $url, StringUtil::specialchars($title), $attributes, Image::getHtml($icon, $label));
+    }
+
+    public function syncMapItemCategoryPivotTable($varValue, $dc)
+    {
+        $this->syncData(deserialize($varValue), 'tl_wem_map_item_category', $dc->id, 'pid', 'category');
+
+        return $varValue;
     }
 }