+++ src/Module/LocationsReader.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * Geodata for Contao Open Source CMS
- * Copyright (c) 2023-2023 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-geodata
@@ -14,14 +14,14 @@ declare(strict_types=1);
 
 namespace WEM\GeoDataBundle\Module;
 
-use Contao\Combiner;
 use Contao\BackendTemplate;
+use Contao\Combiner;
 use Contao\Config;
 use Contao\Environment;
 use Contao\Input;
+use Contao\System;
 use WEM\GeoDataBundle\Model\Map;
 use WEM\GeoDataBundle\Model\MapItem;
-use WEM\GeoDataBundle\Controller\ClassLoader;
 
 /**
  * Front end module "locations reader".
@@ -88,13 +88,12 @@ class LocationsReader extends Core
                 throw new \Exception(sprintf($GLOBALS['TL_LANG']['WEM']['LOCATIONS']['ERROR']['pageNotFound'], Environment::get('uri')));
             }
 
-
             $arrItem = $this->getLocation($objItem);
             $objMap = Map::findByPk($objItem->pid);
             $this->Template->item = $arrItem;
             $this->Template->map = $objMap->row();
             $this->Template->shouldBeIndexed = $objMap->row();
-            
+
             // Load the libraries
             $strVersion = 1;
             $objCssCombiner = new Combiner();
@@ -106,6 +105,23 @@ class LocationsReader extends Core
             $GLOBALS['TL_JAVASCRIPT'][] = 'https://unpkg.com/leaflet-gesture-handling@latest/dist/leaflet-gesture-handling.min.js';
             // And add them to pages
             $GLOBALS['TL_HEAD'][] = sprintf('<link rel="stylesheet" href="%s">', $objCssCombiner->getCombinedFile());
+
+            // Override page details
+            $responseContext = System::getContainer()->get('contao.routing.response_context_accessor')->getResponseContext();
+            if ($responseContext && $responseContext->has(\Contao\CoreBundle\Routing\ResponseContext\HtmlHeadBag\HtmlHeadBag::class)) {
+                /** @var HtmlHeadBag $htmlHeadBag */
+                $htmlHeadBag = $responseContext->get(\Contao\CoreBundle\Routing\ResponseContext\HtmlHeadBag\HtmlHeadBag::class);
+                $htmlHeadBag->setTitle($arrItem['title']);
+
+                $htmlDecoder = System::getContainer()->get('contao.string.html_decoder');
+                $htmlHeadBag->setMetaDescription($htmlDecoder->htmlToPlainText($arrItem['content']));
+            }else{
+                $htmlDecoder = System::getContainer()->get('contao.string.html_decoder');
+                $objPage->ogTitle = $arrItem['title'];
+                $objPage->pageTitle = $arrItem['title'];
+                $objPage->description = $htmlDecoder->htmlToPlainText($arrItem['content']);
+            }
+
         } catch (\Exception $e) {
             $this->Template->error = true;
             $this->Template->msg = $e->getMessage();