+++ src/Resources/contao/dca/tl_wem_map_item.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * Geodata for Contao Open Source CMS
- * Copyright (c) 2023-2023 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-geodata
@@ -14,7 +14,7 @@ declare(strict_types=1);
 
 /*
  * Geodata for Contao Open Source CMS
- * Copyright (c) 2023-2023 Web ex Machina
+ * Copyright (c) 2015-2023 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-geodata
@@ -30,14 +30,11 @@ $GLOBALS['TL_DCA']['tl_wem_map_item'] = [
     'config' => [
         'dataContainer' => 'Table',
         'ptable' => 'tl_wem_map',
-        'ctable' => ['tl_content', 'tl_wem_map_item_attribute_value'],
+        'ctable' => ['tl_content', 'tl_wem_map_item_attribute_value', 'tl_wem_map_item_category'],
         'switchToEdit' => true,
         'enableVersioning' => true,
         'onload_callback' => [
             [\WEM\GeoDataBundle\DataContainer\MapItem::class, 'checkIfGeocodeExists'],
-        ],
-        'onsubmit_callback' => [
-            //array(\WEM\GeoDataBundle\DataContainer\MapItem::class, 'fetchCoordinates'),
         ],
         'sql' => [
             'keys' => [
@@ -131,7 +128,7 @@ $GLOBALS['TL_DCA']['tl_wem_map_item'] = [
     // Palettes
     'palettes' => [
         'default' => '
-            {location_legend},title,alias,category,published, publishedAt, publishedUntil;
+            {location_legend},title,alias,categories,published, publishedAt, publishedUntil;
             {street_legend},country,admin_lvl_1,admin_lvl_2,admin_lvl_3,city,postal,street;
             {coords_legend},lat,lng;
             {data_legend},picture,teaser;
@@ -180,7 +177,7 @@ $GLOBALS['TL_DCA']['tl_wem_map_item'] = [
         'category' => [
             'label' => &$GLOBALS['TL_LANG']['tl_wem_map_item']['category'],
             'exclude' => true,
-            'filter' => true,
+            // 'filter' => true,
             'sorting' => true,
             'flag' => 11,
             'inputType' => 'select',
@@ -189,6 +186,23 @@ $GLOBALS['TL_DCA']['tl_wem_map_item'] = [
             'eval' => ['chosen' => true, 'includeBlankOption' => true, 'tl_class' => 'w50'],
             'sql' => "int(10) unsigned NOT NULL default '0'",
             'relation' => ['type' => 'hasOne', 'load' => 'lazy'],
+        ],
+        'categories' => [
+            'label' => &$GLOBALS['TL_LANG']['tl_wem_map_item']['categories'],
+            'exclude' => true,
+            'filter' => true,
+            'sorting' => true,
+            'flag' => 11,
+            'inputType' => 'select',
+            'foreignKey' => 'tl_wem_map_category.title',
+            'options_callback' => [\WEM\GeoDataBundle\DataContainer\MapItem::class, 'getMapCategories'],
+            'load_callback' => [[\WEM\GeoDataBundle\DataContainer\MapItem::class, 'assignDefaultCategoryIfNew']],
+            'save_callback' => [
+                [\WEM\GeoDataBundle\DataContainer\MapItem::class, 'syncMapItemCategoryPivotTable'],
+            ],
+            'eval' => ['chosen' => true, 'includeBlankOption' => true, 'multiple' => true, 'mandatory' => true, 'tl_class' => 'w50'],
+            'sql' => 'blob NULL',
+            'relation' => ['type' => 'belongsTo', 'load' => 'eager'],
         ],
         'published' => [
             'label' => &$GLOBALS['TL_LANG']['tl_wem_map_item']['published'],