+++ src/Resources/contao/dca/tl_wem_map_item_category.php
@@ -1,0 +1,59 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Geodata for Contao Open Source CMS
+ * Copyright (c) 2015-2023 Web ex Machina
+ *
+ * @category ContaoBundle
+ * @package  Web-Ex-Machina/contao-geodata
+ * @author   Web ex Machina <contact@webexmachina.fr>
+ * @link     https://github.com/Web-Ex-Machina/contao-geodata/
+ */
+
+$GLOBALS['TL_DCA']['tl_wem_map_item_category'] = [
+    // Config
+    'config' => [
+        'dataContainer' => Contao\DC_Table::class,
+        'ptable' => 'tl_wem_map_item',
+        'ctable' => [],
+        'switchToEdit' => true,
+        'enableVersioning' => true,
+        'sql' => [
+            'keys' => [
+                'id' => 'primary',
+                'pid' => 'index',
+            ],
+        ],
+        'ondelete_callback' => [
+            [\WEM\GeoDataBundle\DataContainer\MapItemCategory::class, 'ondeleteCallback'],
+        ],
+    ],
+
+    // Fields
+    'fields' => [
+        'id' => [
+            'sql' => 'int(10) unsigned NOT NULL auto_increment',
+        ],
+        'tstamp' => [
+            'flag' => 8,
+            'sql' => "varchar(10) NOT NULL default ''",
+        ],
+        'created_at' => [
+            'default' => time(),
+            'flag' => 8,
+            'sql' => "varchar(10) NOT NULL default ''",
+        ],
+        'pid' => [
+            'foreignKey' => 'tl_wem_map_item.id',
+            'sql' => "int(10) unsigned NOT NULL default '0'",
+            'relation' => ['type' => 'belongsTo', 'load' => 'eager'],
+        ],
+        'category' => [
+            'foreignKey' => 'tl_wem_map_category.title',
+            'sql' => "int(10) unsigned NOT NULL default '0'",
+            'relation' => ['type' => 'belongsTo', 'load' => 'eager'],
+        ],
+    ],
+];