+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Extension "Geodata" for Contao Open Source CMS
 ========
 
+1.0.0 - 2023-08-07
+---
+First release
+
+1.0.0-rc6 - 2023-07-17
+---
+- UPDATED : map items can now have multiple categories
+- UPDATED : filters in locations' list
+
+1.0.0-rc5 - 2023-02-06
+---
+- UPDATED : moving hook WEMGEODATABUILDFILTERSSINGLEFILTEROPTION calls to allow more flexibility
+
 1.0.0-rc4 - 2023-02-02
 ---
 - UPDATED : translations for administrative levels